fix openID 2.0 directed identity endpoints causing 1.1 verified_identities to fail
_discover_acceptable_endpoints (force_version=>1) wasn't filtering out
openID 2 directed identity endpoints, which, being version 2 would then
be sorted to the front of the @$possible_endpoints list, the rest of the
list would be dropped and then the verified identity would fail because
the endpoint wasn't verion 1.

So we clarify that _discover_acceptable_endpoints (force_version=>n)
ONLY returns endpoints of version n, which appears to be the original
intent, and then the endpoint version check in verified_identity can
go away entirely.